scrap 디자인패턴 정리 완료
